requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-WTF==1.1.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
python-docx==0.8.11
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
